feat: PostCard 버튼 분리
